Created initial pages, components, and types for workout posts

+++ src/routes/login/+page.svelte
@@ -1,12 +1,14 @@
 <script lang="ts">
 	import { goto } from '$app/navigation';
-	import { redirect } from '@sveltejs/kit';
-	import { getAuth, GoogleAuthProvider, EmailAuthProvider } from 'firebase/auth';
-	import * as firebaseui from 'firebaseui';
-	import 'firebaseui/dist/firebaseui.css';
+  import { getAuth, GoogleAuthProvider, EmailAuthProvider } from 'firebase/auth';
+  import * as firebaseui from 'firebaseui'
+  import 'firebaseui/dist/firebaseui.css'
 	import { onDestroy, onMount } from 'svelte';
 
-	let ui;
+  let ui: firebaseui.auth.AuthUI | undefined;
+  
+  onMount(async () => {
+    let auth = getAuth();
 
 	onMount(async () => {
 		let auth = getAuth();
@@ -17,19 +19,21 @@
 		if (auth.currentUser != null) {
 			await goto('/user');
 		}
-
-		var uiConfig = {
-			signInSuccessUrl: '/user',
-			signInOptions: [GoogleAuthProvider.PROVIDER_ID, EmailAuthProvider.PROVIDER_ID]
-		};
+    var uiConfig = {
+      signInSuccessUrl: '/user',
+      signInOptions: [
+        GoogleAuthProvider.PROVIDER_ID,
+      ],
+    };
 
 		ui = new firebaseui.auth.AuthUI(auth);
 		ui.start('#firebaseui-auth-container', uiConfig);
 	});
 
-	onDestroy(() => {
-		ui.delete();
-	});
+  onDestroy(() => {
+    if (ui) ui.delete();
+  })
+
 </script>
 
 <div class="w-full h-full flex flex-col justify-center items-center font-bold text-lg">
